\name{EnsDbLite.Mmusculus.84}
\docType{package}

\alias{EnsDbLite.Mmusculus.84-package}
\alias{EnsDbLite.Mmusculus.84}
\alias{EnsDbLite.Mmusculus.84}


\title{Ensembl-based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Wed Apr 20 15:33:50 2016 and based on the GRCm38
}

\author{.}



\examples{
## load the library
##library(EnsDbLite.Mmusculus.84)
## list the contents that are loaded into memory
ls('package:EnsDbLite.Mmusculus.84')
## show the db object that is loaded by calling it's name
EnsDbLite.Mmusculus.84

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
